Init does not overwrite existing config file.

+++ lib/mir/init.rb
@@ -5,12 +5,15 @@ module Mir
     class << self
       def init_project_dir(dirname)
         config_file = File.join(dirname, '.mir')
-        config = {
-          'email' => 'EMAIL GOES HERE...',
-          'token' => 'TOKEN GOES HERE...'
-        }
 
-        File.write(config_file, config.to_yaml)
+        unless File.exists?(config_file)
+          config = {
+            'email' => 'EMAIL GOES HERE...',
+            'token' => 'TOKEN GOES HERE...'
+          }
+
+          File.write(config_file, config.to_yaml)
+        end
       end
     end
   end
